refactor: moved type-checker errors, and refactor a bit other piece of codes
