+++ crates/checker/src/lib.rs
@@ -421,381 +421,3 @@ impl TypeChecker {
         })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::path::Path;
-
-    use super::TypeChecker;
-    use crate::{
-        checker::error::{TypeError, TypeErrorKind},
-        hir::{
-            ExpressionId, SlynxHir,
-            model::{
-                HirDeclarationKind, HirExpression, HirExpressionKind, HirStatementKind, HirType,
-            },
-        },
-        lexer::Lexer,
-        parser::Parser,
-    };
-
-    fn load_hir_from_source(source: &str) -> SlynxHir {
-        let tokens = Lexer::tokenize(source).expect("source should tokenize");
-        let declarations = Parser::new(tokens)
-            .parse_declarations()
-            .expect("source should parse");
-        let mut hir = SlynxHir::new();
-        hir.generate(declarations).expect("HIR should generate");
-        hir
-    }
-
-    fn load_hir(path: &str) -> SlynxHir {
-        let source_path = Path::new(env!("CARGO_MANIFEST_DIR")).join("..").join(path);
-        let source = std::fs::read_to_string(&source_path).expect("source file should exist");
-        load_hir_from_source(&source)
-    }
-
-    fn main_call_args(hir: &mut SlynxHir) -> Option<&mut Vec<HirExpression>> {
-        for declaration in &mut hir.declarations {
-            let HirDeclarationKind::Function {
-                name, statements, ..
-            } = &mut declaration.kind
-            else {
-                continue;
-            };
-
-            if let Some(main_symbol) = hir.modules.retrieve_symbol("main")
-                && main_symbol == *name
-            {
-                continue;
-            }
-
-            for statement in statements {
-                let expr = match &mut statement.kind {
-                    HirStatementKind::Variable { value, .. } => value,
-                    HirStatementKind::Expression { expr } => expr,
-                    HirStatementKind::Return { expr } => expr,
-                    HirStatementKind::Assign { value, .. } => value,
-                    HirStatementKind::While { .. } => continue,
-                };
-                let HirExpressionKind::FunctionCall { args, .. } = &mut expr.kind else {
-                    continue;
-                };
-                return Some(args);
-            }
-        }
-        None
-    }
-
-    #[test]
-    fn function_calls_work_with_mixed_declaration_order() {
-        let mut hir = load_hir("slynx/functioncall.slynx");
-        TypeChecker::check(&mut hir).expect("function call should resolve with declaration ids");
-    }
-
-    #[test]
-    fn rejects_function_call_with_missing_arg() {
-        let mut hir = load_hir("slynx/functioncall.slynx");
-        let args = main_call_args(&mut hir).expect("expected to find a function call in main");
-        args.pop();
-
-        let err = TypeChecker::check(&mut hir)
-            .expect_err("type checker should reject function calls with missing args");
-        let type_error = err
-            .downcast_ref::<TypeError>()
-            .expect("error should come from type checker");
-
-        match &type_error.kind {
-            TypeErrorKind::InvalidFuncallArgLength {
-                expected_length,
-                received_length,
-            } => {
-                assert_eq!(*expected_length, 2);
-                assert_eq!(*received_length, 1);
-            }
-            other => panic!("expected InvalidFuncallArgLength, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn rejects_function_call_with_extra_arg() {
-        let mut hir = load_hir("slynx/functioncall.slynx");
-        let args = main_call_args(&mut hir).expect("expected to find a function call in main");
-        let template = args.first().expect("call should have at least one arg");
-        args.push(HirExpression {
-            id: ExpressionId::new(),
-            ty: template.ty,
-            kind: HirExpressionKind::Int(0),
-            span: template.span.clone(),
-        });
-
-        let err = TypeChecker::check(&mut hir)
-            .expect_err("type checker should reject function calls with extra args");
-        let type_error = err
-            .downcast_ref::<TypeError>()
-            .expect("error should come from type checker");
-
-        match &type_error.kind {
-            TypeErrorKind::InvalidFuncallArgLength {
-                expected_length,
-                received_length,
-            } => {
-                assert_eq!(*expected_length, 2);
-                assert_eq!(*received_length, 3);
-            }
-            other => panic!("expected InvalidFuncallArgLength, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn rejects_function_call_with_wrong_argument_type() {
-        let mut hir = load_hir("slynx/functioncall.slynx");
-        let bool_ty = hir.bool_type();
-        let args = main_call_args(&mut hir).expect("expected to find a function call in main");
-        let first_arg = args.first_mut().expect("call should have at least one arg");
-        first_arg.kind = HirExpressionKind::Bool(true);
-        first_arg.ty = bool_ty;
-
-        let err = TypeChecker::check(&mut hir)
-            .expect_err("type checker should reject function calls with wrong arg type");
-        let type_error = err
-            .downcast_ref::<TypeError>()
-            .expect("error should come from type checker");
-
-        assert!(
-            matches!(type_error.kind, TypeErrorKind::IncompatibleTypes { .. }),
-            "expected IncompatibleTypes, got {:?}",
-            type_error.kind
-        );
-    }
-
-    #[test]
-
-    fn rejects_function_without_return_value_for_non_void_return_type() {
-        let mut hir = load_hir_from_source(
-            r#"
-            func main(): int {
-                let x = 12;
-            }
-            "#,
-        );
-
-        let err = TypeChecker::check(&mut hir).expect_err("non-void functions must return a value");
-        let type_error = err
-            .downcast_ref::<TypeError>()
-            .expect("error should come from type checker");
-
-        match &type_error.kind {
-            TypeErrorKind::MissingReturnValue { expected } => {
-                assert!(
-                    matches!(expected, HirType::Int),
-                    "expected missing int return, got {expected:?}"
-                );
-            }
-            other => panic!("expected MissingReturnValue, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn preserves_non_expression_tail_statement_in_function_body() {
-        let hir = load_hir_from_source(
-            r#"
-            func main(): void {
-                let x = 12;
-            }
-            "#,
-        );
-        let main_symbol = hir
-            .modules
-            .symbols_resolver
-            .retrieve("main")
-            .copied()
-            .expect("main symbol should exist");
-
-        let main_fn = hir
-            .declarations
-            .iter()
-            .find(|declaration| {
-                matches!(
-                    declaration.kind,
-                    HirDeclarationKind::Function { ref name, .. } if *name == main_symbol
-                )
-            })
-            .expect("main function should exist");
-
-        let HirDeclarationKind::Function { statements, .. } = &main_fn.kind else {
-            unreachable!();
-        };
-
-        assert_eq!(
-            statements.len(),
-            1,
-            "last non-expression statement should be preserved"
-        );
-        assert!(
-            matches!(statements[0].kind, HirStatementKind::Variable { .. }),
-            "expected trailing let statement to stay in the body"
-        );
-    }
-
-    #[test]
-    fn rejects_while_with_non_boolean_condition() {
-        let mut hir = load_hir_from_source(
-            r#"
-            func main(): void {
-                while 10 {
-                    0;
-                }
-            }
-            "#,
-        );
-
-        let err = TypeChecker::check(&mut hir).expect_err("while conditions should require bool");
-        let type_error = err
-            .downcast_ref::<TypeError>()
-            .expect("error should come from type checker");
-
-        assert!(
-            matches!(type_error.kind, TypeErrorKind::IncompatibleTypes { .. }),
-            "expected IncompatibleTypes, got {:?}",
-            type_error.kind
-        );
-    }
-
-    #[test]
-    fn rejects_invalid_statement_inside_while_body() {
-        let mut hir = load_hir_from_source(
-            r#"
-            func takes_int(value: int): void {}
-
-            func main(): void {
-                while true {
-                    takes_int(false);
-                }
-            }
-            "#,
-        );
-
-        let err =
-            TypeChecker::check(&mut hir).expect_err("while body statements should be type-checked");
-        let type_error = err
-            .downcast_ref::<TypeError>()
-            .expect("error should come from type checker");
-
-        assert!(
-            matches!(type_error.kind, TypeErrorKind::IncompatibleTypes { .. }),
-            "expected IncompatibleTypes, got {:?}",
-            type_error.kind
-        );
-    }
-
-    #[test]
-    fn resolves_field_access_for_variables_typed_via_alias() {
-        let mut hir = load_hir_from_source(
-            r#"
-            object Person {
-                age: int,
-            }
-
-            alias PersonAlias = Person;
-
-            func make_person(): PersonAlias {
-                Person(age: 22)
-            }
-
-            func main(): int {
-                let person = make_person();
-                person.age
-            }
-            "#,
-        );
-
-        TypeChecker::check(&mut hir).expect("field access should resolve through aliases");
-    }
-
-    #[test]
-    fn resolves_tuple_access_for_tuple_variables() {
-        let mut hir = load_hir_from_source(
-            r#"
-            func main(): int {
-                let pair = (10, 20);
-                pair.0
-            }
-            "#,
-        );
-
-        TypeChecker::check(&mut hir).expect("tuple access should resolve through the checker");
-    }
-
-    #[test]
-    fn resolves_named_field_access_after_tuple_access() {
-        let mut hir = load_hir_from_source(
-            r#"
-            object Person {
-                age: int,
-            }
-
-            func main(): int {
-                let pair = (Person(age: 22), "ok");
-                pair.0.age
-            }
-            "#,
-        );
-
-        TypeChecker::check(&mut hir)
-            .expect("named field access after tuple access should resolve cleanly");
-    }
-
-    #[test]
-    fn rejects_tuple_access_with_invalid_index() {
-        let mut hir = load_hir_from_source(
-            r#"
-            func main(): int {
-                let pair = (10, 20);
-                pair.2
-            }
-            "#,
-        );
-
-        let err =
-            TypeChecker::check(&mut hir).expect_err("tuple accesses should reject invalid indexes");
-        let type_error = err
-            .downcast_ref::<TypeError>()
-            .expect("error should come from type checker");
-
-        match &type_error.kind {
-            TypeErrorKind::InvalidTupleIndex { index, length } => {
-                assert_eq!(*index, 2);
-                assert_eq!(*length, 2);
-            }
-            other => panic!("expected InvalidTupleIndex, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn rejects_tuple_access_on_non_tuple_values() {
-        let mut hir = load_hir_from_source(
-            r#"
-            func main(): int {
-                let value = 10;
-                value.0
-            }
-            "#,
-        );
-
-        let err =
-            TypeChecker::check(&mut hir).expect_err("non-tuples should reject tuple-style access");
-        let type_error = err
-            .downcast_ref::<TypeError>()
-            .expect("error should come from type checker");
-
-        assert!(
-            matches!(
-                type_error.kind,
-                TypeErrorKind::InvalidTupleAccessTarget { .. }
-            ),
-            "expected InvalidTupleAccessTarget, got {:?}",
-            type_error.kind
-        );
-    }
-}
